A rather large commit that includes a few things that have been lingering for weeks while proposal writing.
Major changes:

1.  Rewrote the subgrid time integrator to be able to handle both "hybridize-then-discretize" and "discretize-then-hybridize".

2.  Updated all of the local basis functions to be consistent with the new requirements in Intrepid2.

Will soon add regression tests to cover the first major change.

The second major change should fix most of the currently failing regression tests.  There are still 2 failing for undetermined reasons.

+++ src/tools/wells.hpp
@@ -97,7 +97,7 @@ namespace MrHyDE {
                      KOKKOS_LAMBDA (const int elem) {
           ScalarT C = std::log(0.25*std::exp(-0.5772)*h(elem)/2.0);
 
-          for (size_t pt=0; pt<numIP; pt++) {
+          for (int pt=0; pt<numIP; pt++) {
             if (wellnum == 0) source_kv(elem,pt) = source(elem,pt); // initialize properly
             // this allows for an additional volumetric source
             // do not name any of your wells the same as the default volumetric source
